cleanup per pr comments
