+++ crates/sui-graphql-rpc/Cargo.toml
@@ -39,6 +39,7 @@ uuid.workspace = true
 # Used for dummy data
 bcs.workspace = true
 sui-sdk.workspace = true
+sui-json-rpc.workspace = true
 sui-json-rpc-types.workspace = true
 sui-indexer.workspace = true
 sui-types.workspace = true
