+++ crates/sui-graphql-rpc/src/types/address.rs
@@ -2,9 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use async_graphql::{connection::Connection, *};
+use sui_json_rpc::name_service::NameServiceConfig;
 
 use crate::context_data::db_data_provider::PgManager;
-use crate::context_data::name_service::NameServiceConfig;
 
 use super::{
     balance::Balance,
